+++ src/achievements/achievement.hh
@@ -30,7 +30,7 @@ namespace shiro::achievements {
         virtual ~achievement() = default;
 
         virtual void verify(
-            std::vector<std::pair<size_t, std::string>>& output,
+            std::vector<std::pair<int64_t, std::string>>& output,
             const std::shared_ptr<shiro::users::user>& user, 
             const shiro::beatmaps::beatmap& beatmap, 
             const shiro::scores::score& score
