+++ src/achievements/impl/skill/combo.cc
@@ -32,7 +32,7 @@ namespace keys {
 }
 
 void shiro::achievements::combo::verify(
-    std::vector<std::pair<size_t, std::string>>& output,
+    std::vector<std::pair<int64_t, std::string>>& output,
     const std::shared_ptr<shiro::users::user>& user, 
     const shiro::beatmaps::beatmap& beatmap, 
     const shiro::scores::score& score
